+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to Family Diagram are documented in this file.
 
+## 2.1.17
+
+## 2.1.17b1
+
+- Click row in timeline sets the diagram to that date.
+- Fixed broked "Add birth" and "Edit birth" buttons in person properties
+- Fixed person picker in add form not showing renamed people in auto-complete.
+- Can now use down/up arrows + enter/return to pick a person via autocomplete in add event form.
+- Visual timeline shows person names now.
+
 ## 2.1.16
 
 ## 2.1.16b2